fix: several issues on error propagation and persistence feat: remove settlement/debt ratio validation

+++ internal/domain/users.go
@@ -17,7 +17,7 @@ type (
 		IsEmailVerified bool
 		LastName        string
 		PasswordHash    string
-		LedgersCount    int64
+		LedgersCount    int32
 	}
 
 	NewUserRequest struct {
